sales_forecasting_mlops/requirements.txt
pyspark                      
pandas               
pyarrow            
numpy               
scikit-learn          
xgboost       
snowflake-connector-python   
python-dotenv        
pyyaml          
matplotlib  
streamlit      